+++ projection/winrt-Windows.Devices.Enumeration.Pnp/py.Windows.Devices.Enumeration.Pnp.cpp
@@ -599,7 +599,7 @@ namespace py::cpp::Windows::Devices::Enumeration::Pnp
 
             auto count = self->obj.GetMany(static_cast<uint32_t>(start), items);
 
-            if (count != length)
+            if (count != static_cast<uint32_t>(length))
             {
                 PyErr_Format(
                     PyExc_RuntimeError,
