+++ projection/winrt-Windows.Devices.Lights.Effects/py.Windows.Devices.Lights.Effects.cpp
@@ -2322,7 +2322,7 @@ namespace py::cpp::Windows::Devices::Lights::Effects
 
             auto count = self->obj.GetMany(static_cast<uint32_t>(start), items);
 
-            if (count != length)
+            if (count != static_cast<uint32_t>(length))
             {
                 PyErr_Format(
                     PyExc_RuntimeError,
