+++ projection/winrt-Windows.Globalization.Collation/py.Windows.Globalization.Collation.cpp
@@ -470,7 +470,7 @@ namespace py::cpp::Windows::Globalization::Collation
 
             auto count = self->obj.GetMany(static_cast<uint32_t>(start), items);
 
-            if (count != length)
+            if (count != static_cast<uint32_t>(length))
             {
                 PyErr_Format(
                     PyExc_RuntimeError,
