+++ projection/winrt-Windows.UI.Composition.Scenes/py.Windows.UI.Composition.Scenes.cpp
@@ -802,7 +802,7 @@ namespace py::cpp::Windows::UI::Composition::Scenes
 
             auto count = self->obj.GetMany(static_cast<uint32_t>(start), items);
 
-            if (count != length)
+            if (count != static_cast<uint32_t>(length))
             {
                 PyErr_Format(
                     PyExc_RuntimeError,
@@ -3296,7 +3296,7 @@ namespace py::cpp::Windows::UI::Composition::Scenes
 
             auto count = self->obj.GetMany(static_cast<uint32_t>(start), items);
 
-            if (count != length)
+            if (count != static_cast<uint32_t>(length))
             {
                 PyErr_Format(
                     PyExc_RuntimeError,
