+++ projection/winrt-Windows.UI.Composition/py.Windows.UI.Composition.cpp
@@ -4223,7 +4223,7 @@ namespace py::cpp::Windows::UI::Composition
 
             auto count = self->obj.GetMany(static_cast<uint32_t>(start), items);
 
-            if (count != length)
+            if (count != static_cast<uint32_t>(length))
             {
                 PyErr_Format(
                     PyExc_RuntimeError,
@@ -13141,7 +13141,7 @@ namespace py::cpp::Windows::UI::Composition
 
             auto count = self->obj.GetMany(static_cast<uint32_t>(start), items);
 
-            if (count != length)
+            if (count != static_cast<uint32_t>(length))
             {
                 PyErr_Format(
                     PyExc_RuntimeError,
@@ -14375,7 +14375,7 @@ namespace py::cpp::Windows::UI::Composition
 
             auto count = self->obj.GetMany(static_cast<uint32_t>(start), items);
 
-            if (count != length)
+            if (count != static_cast<uint32_t>(length))
             {
                 PyErr_Format(
                     PyExc_RuntimeError,
