+++ projection/winrt-Windows.UI.Xaml.Documents/py.Windows.UI.Xaml.Documents.cpp
@@ -973,7 +973,7 @@ namespace py::cpp::Windows::UI::Xaml::Documents
 
             auto count = self->obj.GetMany(static_cast<uint32_t>(start), items);
 
-            if (count != length)
+            if (count != static_cast<uint32_t>(length))
             {
                 PyErr_Format(
                     PyExc_RuntimeError,
@@ -3250,7 +3250,7 @@ namespace py::cpp::Windows::UI::Xaml::Documents
 
             auto count = self->obj.GetMany(static_cast<uint32_t>(start), items);
 
-            if (count != length)
+            if (count != static_cast<uint32_t>(length))
             {
                 PyErr_Format(
                     PyExc_RuntimeError,
@@ -5998,7 +5998,7 @@ namespace py::cpp::Windows::UI::Xaml::Documents
 
             auto count = self->obj.GetMany(static_cast<uint32_t>(start), items);
 
-            if (count != length)
+            if (count != static_cast<uint32_t>(length))
             {
                 PyErr_Format(
                     PyExc_RuntimeError,
