+++ projection/winrt-Windows.UI.Xaml.Media.Animation/py.Windows.UI.Xaml.Media.Animation.cpp
@@ -2319,7 +2319,7 @@ namespace py::cpp::Windows::UI::Xaml::Media::Animation
 
             auto count = self->obj.GetMany(static_cast<uint32_t>(start), items);
 
-            if (count != length)
+            if (count != static_cast<uint32_t>(length))
             {
                 PyErr_Format(
                     PyExc_RuntimeError,
@@ -5985,7 +5985,7 @@ namespace py::cpp::Windows::UI::Xaml::Media::Animation
 
             auto count = self->obj.GetMany(static_cast<uint32_t>(start), items);
 
-            if (count != length)
+            if (count != static_cast<uint32_t>(length))
             {
                 PyErr_Format(
                     PyExc_RuntimeError,
@@ -11790,7 +11790,7 @@ namespace py::cpp::Windows::UI::Xaml::Media::Animation
 
             auto count = self->obj.GetMany(static_cast<uint32_t>(start), items);
 
-            if (count != length)
+            if (count != static_cast<uint32_t>(length))
             {
                 PyErr_Format(
                     PyExc_RuntimeError,
@@ -13483,7 +13483,7 @@ namespace py::cpp::Windows::UI::Xaml::Media::Animation
 
             auto count = self->obj.GetMany(static_cast<uint32_t>(start), items);
 
-            if (count != length)
+            if (count != static_cast<uint32_t>(length))
             {
                 PyErr_Format(
                     PyExc_RuntimeError,
@@ -20976,7 +20976,7 @@ namespace py::cpp::Windows::UI::Xaml::Media::Animation
 
             auto count = self->obj.GetMany(static_cast<uint32_t>(start), items);
 
-            if (count != length)
+            if (count != static_cast<uint32_t>(length))
             {
                 PyErr_Format(
                     PyExc_RuntimeError,
@@ -21684,7 +21684,7 @@ namespace py::cpp::Windows::UI::Xaml::Media::Animation
 
             auto count = self->obj.GetMany(static_cast<uint32_t>(start), items);
 
-            if (count != length)
+            if (count != static_cast<uint32_t>(length))
             {
                 PyErr_Format(
                     PyExc_RuntimeError,
