+++ projection/winrt-Windows.UI.Xaml/py.Windows.UI.Xaml.cpp
@@ -5823,7 +5823,7 @@ namespace py::cpp::Windows::UI::Xaml
 
             auto count = self->obj.GetMany(static_cast<uint32_t>(start), items);
 
-            if (count != length)
+            if (count != static_cast<uint32_t>(length))
             {
                 PyErr_Format(
                     PyExc_RuntimeError,
@@ -15336,7 +15336,7 @@ namespace py::cpp::Windows::UI::Xaml
 
             auto count = self->obj.GetMany(static_cast<uint32_t>(start), items);
 
-            if (count != length)
+            if (count != static_cast<uint32_t>(length))
             {
                 PyErr_Format(
                     PyExc_RuntimeError,
@@ -17253,7 +17253,7 @@ namespace py::cpp::Windows::UI::Xaml
 
             auto count = self->obj.GetMany(static_cast<uint32_t>(start), items);
 
-            if (count != length)
+            if (count != static_cast<uint32_t>(length))
             {
                 PyErr_Format(
                     PyExc_RuntimeError,
@@ -17939,7 +17939,7 @@ namespace py::cpp::Windows::UI::Xaml
 
             auto count = self->obj.GetMany(static_cast<uint32_t>(start), items);
 
-            if (count != length)
+            if (count != static_cast<uint32_t>(length))
             {
                 PyErr_Format(
                     PyExc_RuntimeError,
@@ -25392,7 +25392,7 @@ namespace py::cpp::Windows::UI::Xaml
 
             auto count = self->obj.GetMany(static_cast<uint32_t>(start), items);
 
-            if (count != length)
+            if (count != static_cast<uint32_t>(length))
             {
                 PyErr_Format(
                     PyExc_RuntimeError,
