+++ projection/winrt-sdk/src/winrt_sdk/pywinrt/py.Windows.Foundation.Collections.h
@@ -1853,7 +1853,7 @@ namespace py::impl::Windows::Foundation::Collections
 
                 auto count = _obj.GetMany(static_cast<uint32_t>(start), items);
 
-                if (count != length)
+                if (count != static_cast<uint32_t>(length))
                 {
                     PyErr_Format(
                         PyExc_RuntimeError,
@@ -2124,7 +2124,7 @@ namespace py::impl::Windows::Foundation::Collections
 
                 auto count = _obj.GetMany(static_cast<uint32_t>(start), items);
 
-                if (count != length)
+                if (count != static_cast<uint32_t>(length))
                 {
                     PyErr_Format(
                         PyExc_RuntimeError,
@@ -2624,7 +2624,7 @@ namespace py::impl::Windows::Foundation::Collections
 
                 auto count = _obj.GetMany(static_cast<uint32_t>(start), items);
 
-                if (count != length)
+                if (count != static_cast<uint32_t>(length))
                 {
                     PyErr_Format(
                         PyExc_RuntimeError,
